Create TodoList component and import it into App.jsx file
